restructured App class, so that new CoreSeach class contains search logic
